internal/apmgodog: move code to to _test package
